% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.combos}
\alias{.combos}
\title{Number of CVR combinations given a number of regions to choose from
It considers A/B equivalent to B/A.}
\usage{
.combos(x)
}
\description{
Number of CVR combinations given a number of regions to choose from
It considers A/B equivalent to B/A.
}
